pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipery-tooling"
version = "0.1.0"
description = "Internal Pipery tooling for GitHub Action scaffolding and management."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Pipery" }
]
dependencies = ["pyyaml>=6"]

[project.scripts]
pipery-actions = "pipery_tooling.cli:main"
pipery-steps = "pipery_tooling.steps.cli:main"

[dependency-groups]
dev = ["pytest>=8"]

[tool.hatch.build.targets.wheel]
packages = ["src/pipery_tooling"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_classes = ["*Tests"]
